requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
slack-sdk==3.24.0
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.8.0
google-generativeai==0.3.0
cohere==4.4.6
transformers==4.36.2
azure-ai-ml==1.20.0
boto3==1.29.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
argparse==1.4.0
